Added Location class.
Followed the model in "Duplication of Knowledge about Topology" in "4 rules of simple design" (p. 19-22) to introduce a new Location class that reifies the idea of a location on the grid.

Note that something like this would be "forced" by the constraint that you pass no primitives in as arguments to a function.