Added Location#neighbors
Added an implementation of the neighbors method to Location, and associated tests. I'm not entirely sold on my nested loop solution to neighbors, but I'm not sure of a better approach.

I needed to add an Location#== method so that tests would work. I suspect that if we tried to scale this we'd find that we'd create vast numbers of Location objects, many of which have the same coordinates. If this is supported by profiling, we might want to have a factory method that checks whether we've already creating that Location, and return that instance if we have.