Added Cell and tests for alive_in_next_generation
I made heavy use of mocking in the tests for whether a cell will be alive in the next generation. This allowed me to completely test that logic without getting muddied up in any of the issues of computing how many neighbors a cell has. This is particularly nice because it allowed to do the tests without having to set up a world, create neighborhoods, or any of that complicated and (for these tests) irrelevant stuff.