Split Cell into LivingCell and DeadCell
This seemed like an odd choice when I saw it hinted at early in the book, but his discussion in "Procedural Polymorphism" (p. 37-41) of the value of polymorphism and its usefulness here was quite compelling, and it turns out that this refactoring definitely cleaned up both the test and the production code in very nice ways.