Inlined the alive_in_next_generation methods
The alive_in_next_generation methods in LivingCell and DeadCell were now just one-line calls to one-line methods that really didn't need to be kept separate anymore. I also found naming the helper methods awkward at the time, which was a sign that getting rid of them was probably a reasonable thing.